resolving issue with transfer files

+++ src/java/entities/ResultadosGenerales.java
@@ -14,7 +14,7 @@ public class ResultadosGenerales {
     private Cita cita;
     private String zonaCerebral;
     private String tipoOndaDominante;
-    private float porcentajeTipoOnda;
+    private double porcentajeTipoOnda;
     
     public ResultadosGenerales(){}
     
@@ -34,7 +34,7 @@ public class ResultadosGenerales {
         return this.tipoOndaDominante;
     }
     
-    public float getPorcentajeTipoOnda(){
+    public double getPorcentajeTipoOnda(){
         return this.porcentajeTipoOnda;
     }
     
@@ -74,7 +74,7 @@ public class ResultadosGenerales {
      *
      * @param porcentajeTipoOnda
      */
-    public void setPorcentajeTipoOnda(float porcentajeTipoOnda){
+    public void setPorcentajeTipoOnda(double porcentajeTipoOnda){
         this.porcentajeTipoOnda = porcentajeTipoOnda;
     }
 }
